requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
click==8.1.7
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
flask==2.3.3
flask-cors==6.0.1
requests==2.31.0
